+++ library/cartridge_validate_config.py
@@ -47,6 +47,7 @@ FALOVER_MODES = [
 STATEFUL_FAILOVER_PARAMS = [
     'state_provider',
     'stateboard_params',
+    'etcd2_params',
 ]
 
 STATEFUL_FAILOVER_REQUIRED_PARAMS = [
@@ -55,11 +56,16 @@ STATEFUL_FAILOVER_REQUIRED_PARAMS = [
 
 STATEFUL_FAILOVER_STATE_PROVIDERS = [
     'stateboard',
+    'etcd2',
 ]
 
 STATEBOARD_PROVIDER_REQUIRED_PARAMS = [
     'uri',
     'password',
+]
+
+ETCD2_PROVIDER_REQUIRED_PARAMS = [
+    'prefix',
 ]
 
 STATEBOARD_CONFIG_REQUIRED_PARAMS = [
@@ -150,7 +156,14 @@ def validate_types(vars):
             'stateboard_params': {
                 'uri': str,
                 'password': str
-            }
+            },
+            'etcd2_params': {
+                'prefix': str,
+                'lock_delay': int,
+                'endpoints': [str],
+                'username': str,
+                'password': str,
+            },
         }
     }
 
@@ -390,6 +403,13 @@ def check_failover(found_common_params):
                     errmsg = 'Stateboard password cannot contain symbols other than [a-zA-Z0-9_.~-] ' + \
                         '("{}" found)'.format(m.group())
                     return errmsg
+
+            elif cartridge_failover_params['state_provider'] == 'etcd2':
+                etcd2_params = cartridge_failover_params.get('etcd2_params')
+                if etcd2_params is not None and etcd2_params.get('endpoints') is not None:
+                    for endpoint in etcd2_params['endpoints']:
+                        if not is_valid_advertise_uri(endpoint):
+                            return 'etcd2 endpoints must be specified as "<host>:<port>"'
 
     return None